Fixes with firefox 23 related style issues
Tested with chrome (latest/dev builds), firefox(latest), internet
explorer (9,10/latest). There's a bug with the search in Internet
Explorer 10, IE is a horrible browser and microsoft should doom it to
hell but because we live in this world, we developers must be abused by
IE's requirements.

The fix is the btn-group heights; firefox reduces the heights of input
fields and since the btn-group was a mix of <a> and <input> you can see
where the problem is. The solution was changing both elements to
buttons.